--- a/pandas/core/internals/managers.py
+++ b/pandas/core/internals/managers.py
@@ -1425,7 +1425,14 @@ class BlockManager(libinternals.BlockManager, BaseBlockManager):
         new_columns = unstacker.get_new_columns(self.items)
         new_index = unstacker.new_index
 
-        allow_fill = not unstacker.mask.all()
+        allow_fill = not unstacker.mask_all
+        if allow_fill:
+            # calculating the full mask once and passing it to Block._unstack is
+            #  faster than letting calculating it in each repeated call
+            new_mask2D = (~unstacker.mask).reshape(*unstacker.full_shape)
+            needs_masking = new_mask2D.any(axis=0)
+        else:
+            needs_masking = None
 
         new_blocks: list[Block] = []
         columns_mask: list[np.ndarray] = []
@@ -1446,6 +1453,7 @@ class BlockManager(libinternals.BlockManager, BaseBlockManager):
                 fill_value,
                 new_placement=new_placement,
                 allow_fill=allow_fill,
+                needs_masking=needs_masking,
             )
 
             new_blocks.extend(blocks)