```
from asv_bench.benchmarks.reshape import *
self = Unstack()
self.setup("category")

%timeit self.time_without_last_row("category")
77.3 ms ± 1.97 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- master
21.2 ms ± 788 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- PR
```